Updated test for User model, moved STATES into a constants folder
