+++ app/models/constants/states.rb
